Create itShouldBeAbleToCalculateFlatCost Test

+++ Code/Models/Flat.cs
@@ -3,7 +3,7 @@ namespace Code.Models
     public class Flat
     {
         public int Id { get; set; }
-        public int SquareMeters { get; set; }
+        public float SquareMeters { get; set; }
         public int NumberOfRooms { get; set; }
 
         public int OwnerId { get; set;}
